enable stuff back
