+++ core/store/src/flat/manager.rs
@@ -96,7 +96,7 @@ impl FlatStorageManager {
     ) -> Result<(), StorageError> {
         if let Some(flat_storage) = self.get_flat_storage_for_shard(shard_uid) {
             // Try to update flat head.
-            flat_storage.update_flat_head_impl(&new_flat_head, false).unwrap_or_else(|err| {
+            flat_storage.update_flat_head(&new_flat_head).unwrap_or_else(|err| {
                 match &err {
                     FlatStorageError::BlockNotSupported(_) => {
                         // It's possible that new head is not a child of current flat head, e.g. when we have a
